investigation/pyproject.toml
[tool.poetry]
name = "investigation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
motor = "^3.3.2"
uvicorn = "^0.26.0"
orjson = "^3.9.10"
locust = "^2.20.1"
python-logstash = "^0.4.8"
pyyaml = "^6.0.1"
python-json-logger = "^2.0.7"
psycopg2-binary = "^2.9.9"
pytest = "^7.4.4"
asyncio = "^3.4.3"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.39.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
